+++ prog/attractor/attractor.c
@@ -83,7 +83,7 @@ void fade(image.rgba_t *c) {
 image.rgba_t create_hue(double h) {
     h *= 6.0;
 	int inth = (int)h;
-    double hf = h - inth;
+    double hf = h - (double) inth;
     image.rgba_t r = {};
     switch (inth % 6) {
         case 0: { r.red = 255; r.green = 255 * hf; r.blue = 0; }
